Minor fixes

+++ core/src/main/scala/kafka/server/ConfigHandler.scala
@@ -71,12 +71,12 @@ class ClientIdConfigHandler(private val quotaManagers: Map[Short, ClientQuotaMan
   def processConfigChanges(clientId : String, clientConfig : Properties) = {
     if (clientConfig.containsKey(ClientConfigOverride.ProducerOverride)) {
       quotaManagers.get(RequestKeys.ProduceKey).get.updateQuota(clientId,
-        new Quota(clientConfig.getProperty(ClientConfigOverride.ProducerOverride).toInt, true))
+        new Quota(clientConfig.getProperty(ClientConfigOverride.ProducerOverride).toLong, true))
     }
 
     if (clientConfig.containsKey(ClientConfigOverride.ConsumerOverride)) {
       quotaManagers.get(RequestKeys.FetchKey).get.updateQuota(clientId,
-        new Quota(clientConfig.getProperty(ClientConfigOverride.ConsumerOverride).toInt, true))
+        new Quota(clientConfig.getProperty(ClientConfigOverride.ConsumerOverride).toLong, true))
     }
   }
 }
